LayoutTests/imported/w3c/web-platform-tests/css/motion/offset-path-serialization-expected.txt


FAIL Motion Path Module Level 1: path serialization assert_equals: expected (string) "path(\"M 10 20 Q 40 80 50 70 Q 150 140 140 150\")" but got (undefined) undefined
